Implemented Logout Endpoint
The auth/logout endpoint has been implemented. If a user is not authorized the page returns 401. To do this we had to modify The ConfigureServices function in Startup.cs.

Integration test written which ensures that both the logout feature removes the authenitication cookie and the correct status code is returned if the user is not authorized.